refactor: split plugin internals into modules
Break up the oversized index entry into focused modules for plugin setup and transform flow.

This keeps behavior and public exports the same while making maintenance easier.
