+++ src/constants.ts
@@ -1,0 +1,14 @@
+// Custom extensions to include .vue files
+export const DEFAULT_EXTENSION = [
+  '.js',
+  '.cjs',
+  '.mjs',
+  '.ts',
+  '.tsx',
+  '.jsx',
+  '.vue',
+];
+export const COVERAGE_PUBLIC_PATH = '/__coverage__';
+export const PLUGIN_NAME = 'vite:istanbul';
+export const MODULE_PREFIX = '/@modules/';
+export const NULL_STRING = '\0';
